Create Get-TenantID
This function retrieves the Azure AD tenant ID associated with a specified domain by querying the Azure AD  well-known OpenID configuration endpoint. It extracts the tenant ID from the issuer URL provided in the response.
This function is useful for scenarios where you need to quickly or programmatically determine the tenant ID based on a domain.